testing code

+++ model/response/xrp/account_info.go
@@ -1,6 +1,6 @@
 package xrp
 
 type AccountInfo struct {
-	LedgerCurrentIndex uint32
-	Sequence           uint32
+	LedgerCurrentIndex uint64
+	Sequence           uint64
 }
